requirements.txt
requests
loguru
discord.py[voice]
python-dotenv